+++ python/adb_auto_player/argparse_formatter_factory.py
@@ -56,19 +56,34 @@ def build_argparse_formatter(commands_by_category: dict[str, list[Command]]):
 
         def _format_action(self, action):
             if action.dest == "command":
-                parts = ["\nGame Commands:"]
+                parts = []
 
-                for group_name, group_cmds in sorted(
-                    commands_by_category.items(),
-                    key=lambda item: (item[0].lower() == "generic", item[0].lower()),
-                ):
-                    parts.append(f"- {group_name}:")
-                    for cmd in sorted(group_cmds, key=lambda c: c.name.lower()):
+                common_cmds = commands_by_category.get("Commands", [])
+                if common_cmds:
+                    parts.append("\nCommon Commands:")
+                    for cmd in sorted(common_cmds, key=lambda c: c.name.lower()):
                         tooltip = getattr(cmd.menu_option, "tooltip", "")
                         if tooltip:
                             parts.append(f"    {cmd.name:<30} {tooltip}")
                         else:
                             parts.append(f"    {cmd.name}")
+
+                other_groups = {
+                    k: v for k, v in commands_by_category.items() if k != "Commands"
+                }
+                if other_groups:
+                    parts.append("\nGame Commands:")
+                    for group_name, group_cmds in sorted(
+                        other_groups.items(), key=lambda item: item[0].lower()
+                    ):
+                        parts.append(f"- {group_name}:")
+                        for cmd in sorted(group_cmds, key=lambda c: c.name.lower()):
+                            tooltip = getattr(cmd.menu_option, "tooltip", "")
+                            if tooltip:
+                                parts.append(f"    {cmd.name:<30} {tooltip}")
+                            else:
+                                parts.append(f"    {cmd.name}")
+
                 return "\n".join(parts) + "\n"
 
             return super()._format_action(action)
