+++ python/adb_auto_player/commands/gui_categories.py
@@ -1,0 +1,9 @@
+"""Enum for Command Categories."""
+
+from enum import StrEnum
+
+
+class CommandCategory(StrEnum):
+    """Enumeration for categories used in the GUIs accordion menu."""
+
+    SETTINGS_PHONE_DEBUG = "Settings, Phone & Debug"
